WIP: Bluetooth: Add broadcasting support on BR/EDR
This will be possible on BR/EDR only controllers.

To enable broadcasting in this kind of transport ISCAN must be performed. It
means that BROADCASTER calls the same HCI command than DISCOVERABLE and
CONNECTABLE settings handler. Below is explained the relation among these
settings:

ISCAN will be performed when BROADCASTER or DISCOVERABLE is set.
PSCAN will be performed when CONNECTABLE is set.

DISCOVERABLE still depends on CONNECTABLE.
BROADCASTER doesn't depend neither on DISCOVERABLE nor on CONNECTABLE.